techDebt(buRefactorTraining): Move validateTrainingCsv into training folder
